+++ src/execution/translation/automatic-translate/automatic-translators/DebugAutomaticTranslator.ts
@@ -1,11 +1,11 @@
 import chalk from 'chalk';
-import { IAutomaticTranslator } from './IAutomaticTranslator';
+import { AutomaticTranslator } from './AutomaticTranslator';
 
 /**
  * This will wrap an automatic translator and log each translation into the console
  */
-export class DebugAutomaticTranslator implements IAutomaticTranslator {
-    public constructor(private readonly automaticTranslator: IAutomaticTranslator) {}
+export class DebugAutomaticTranslator implements AutomaticTranslator {
+    public constructor(private readonly automaticTranslator: AutomaticTranslator) {}
 
     public async translate(message: string): Promise<string> {
         const messageTranslated = await this.automaticTranslator.translate(message);