+++ src/execution/translation/automatic-translate/automatic-translators/IAutomaticTranslator.ts
@@ -1,5 +1,5 @@
 import { Promisable } from 'type-fest';
 
-export interface IAutomaticTranslator {
+export interface AutomaticTranslator {
     translate(message: string): Promisable<string>;
 }