+++ src/execution/translation/automatic-translate/automatic-translators/ITranslatorOptions.ts
@@ -1,4 +1,4 @@
-export interface ITranslatorOptions {
+export interface TranslatorOptions {
     from: string;
     to: string;
 }